Added Guards
If the game is to have some form of enemy to defend against, there needs
to be something to defend with. This is where the guards come in.

Guards are different to the existing two types in that they spawn in
pairs, although the two guards from a cell only eat one food between the
two of them. If one or more guards is killed, which they will be, the
nest will naturally produce a new one every couple seconds until the
right number are buzzing about again. As such, having more guard nests
is useful in that you can rebuild your defenses faster.

The weakest enemy, the wasp, can easily be fended off by a pair of
guards. However, stronger enemies will likely take a lot of stinging to
bring down. Thankfully, guards are more than happy to throw themselves
at enemies.

There are two types of enemy, small and large (distinguished by
gameObject tags). Small enemies have a 1/3 chance of killing a guard
when it stings them, large ones are guaranteed to take the guard with
them.